third_party/blink/web_tests/http/tests/misc/performance-entry-serializer.html

<!DOCTYPE html>
<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>
<script>


function validate_object(input) {
  var entry = input;
  var serialized = JSON.parse(JSON.stringify(input));
  for (var attr in entry) {
      if (typeof entry[attr] != 'function' &&
        // - PerformanceEntry.source is a Window object which is not
        //   serializable.
        // - Performance{Mark|Measure}.detail may be a live object
        !(entry[attr] instanceof Window) && attr !== 'detail') {
          var assertMethod = assert_equals
          if (Array.isArray(serialized[attr]) && Array.isArray(entry[attr])) {
            assertMethod = assert_array_equals
          } else if (serialized[attr] instanceof Object && entry[attr] instanceof Object) {
            return validate_object(entry[attr])
          }
          assertMethod(serialized[attr], entry[attr], entry.constructor.name + '.' + attr);
      }
  }
}

test(function() {
    window.performance.mark('mark1');
    window.performance.mark('mark2');
    window.performance.measure('measure1', 'mark1', 'mark2');

    var entries = window.performance.getEntries();
    for (var i = 0; i < entries.length; i++) {
      validate_object(entries[i])
    }
}, 'PerformanceEntry subclasses should serialize all attributes');

</script>
